requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
sentence-transformers
chromadb
pymupdf
python-dotenv
